Move TeleportToStartingPoint out of Outdoor.cpp
It's called from both the indoor and the outdoor loading paths, and
TeleportPoint.cpp is the level-neutral teleport home. uLevel_StartingPointType
moves along with it.
