+++ src/Engine/Graphics/Indoor.cpp
@@ -30,6 +30,7 @@
 #include "Engine/Tables/ItemTable.h"
 #include "Engine/OurMath.h"
 #include "Engine/Party.h"
+#include "Engine/TeleportPoint.h"
 #include "Engine/Snapshots/CompositeSnapshots.h"
 #include "Engine/SpellFxRenderer.h"
 #include "Engine/Time/Timer.h"
